requirements.txt
# DevOnboarder Production Dependencies
# Generated from pyproject.toml project.dependencies
# See pyproject.toml for version constraints and optional dependencies

# Web Framework
fastapi
uvicorn

# HTTP Client
httpx<0.29

# Database
SQLAlchemy<3.0
alembic
psycopg2-binary

# Authentication & Security
passlib[bcrypt]
bcrypt>=4.3.0
PyJWT

# Configuration & Utilities
python-dotenv
pyyaml>=6.0

# Observability
logfire>=4.11.0